+++ src/main/java/com/google/cloud/spanner/pgadapter/ConnectionHandler.java
@@ -29,6 +29,7 @@ import com.google.cloud.spanner.Spanner;
 import com.google.cloud.spanner.SpannerException;
 import com.google.cloud.spanner.SpannerException.ResourceNotFoundException;
 import com.google.cloud.spanner.SpannerExceptionFactory;
+import com.google.cloud.spanner.SpannerOptions;
 import com.google.cloud.spanner.Statement;
 import com.google.cloud.spanner.connection.AbstractStatementParser.StatementType;
 import com.google.cloud.spanner.connection.Connection;
@@ -209,6 +210,7 @@ public class ConnectionHandler implements Runnable {
           connectionOptionsBuilder.setSessionPoolOptions(options.getSessionPoolOptions());
     }
     if (options.isEnableOpenTelemetryMetrics()) {
+      SpannerOptions.enableOpenTelemetryMetrics();
       connectionOptionsBuilder =
           connectionOptionsBuilder.setOpenTelemetry(server.getOpenTelemetry());
     }